+++ packages/gov-ui-kit/src/components/checkbox/index.ts
@@ -1,2 +1,3 @@
 export * from './checkbox';
+export * from './checkboxCard';
 export * from './checkboxGroup';